form を LotteryRoulette の中に移動
